Done more cleaning and addressing the reviews
